refactor: Consolidate AppLoggerRepository in domain module
Move shared AppLoggerRepository interface (log + countKey) to
domain/src/commonMain/. Rename androidApp's interface to
AndroidAppLoggerRepository (adds writeCsvToUri). Delete
AppLoggerDataSource from data module (superseded).

All callers that only need log() now use the domain interface.
Only AppLoggerViewModel uses AndroidAppLoggerRepository for CSV.
