--- AndroidGarage/data/src/commonMain/kotlin/com/chriscartland/garage/data/AppLoggerDataSource.kt
@@ -1,16 +1,0 @@
-package com.chriscartland.garage.data
-
-import kotlinx.coroutines.flow.Flow
-
-/**
- * Shared interface for application event logging.
- *
- * Platform-specific implementations handle storage (Room on Android,
- * CoreData on iOS). The Android implementation adds writeCsvToUri()
- * as an extension — that method is not part of the shared contract.
- */
-interface AppLoggerDataSource {
-    suspend fun log(key: String)
-
-    fun countKey(key: String): Flow<Long>
-}
